+++ analyzer/analyzer.py
@@ -4,10 +4,11 @@ import os
 from tools import logger as logger
 from .container import DataContainer
 from .ards_catboost_dynamic import ARDSCatboostRegressionAnalyzer
-from .ards_explore import ArdsFeatureExplorer
+from .dataset_explore.ards_explore import ArdsFeatureExplorer
 from .ards_nearest_cls import ArdsNearest4ClsAnalyzer
 from .cross_validation import CV_Analyzer
-from .raw_explore import RawFeatureExplorer
+from .dataset_explore.raw_explore import RawFeatureExplorer
+from .vent_catboost_dynamic import VentCatboostDynamicAnalyzer
 
 
 class Analyzer:
@@ -21,6 +22,7 @@ class Analyzer:
             'ards_catboost_dynamic': ARDSCatboostRegressionAnalyzer,
             'ards_feature_explore': ArdsFeatureExplorer,
 
+            'vent_catboost_dynamic': VentCatboostDynamicAnalyzer,
             'cross_validation': CV_Analyzer,
 
             'raw_feature_explore': RawFeatureExplorer,