+++ analyzer/analyzer_utils.py
@@ -1,0 +1,20 @@
+from tools import logger as logger
+import pickle
+import numpy as np
+import tools
+import os
+
+def create_final_result(out_dir):
+    '''Collect result.log from each folder, merge to final result.log'''
+    logger.info('Creating final result')
+    with open(os.path.join(out_dir, 'final_result.log'), 'w') as final_f:
+        for dir in os.listdir(out_dir):
+            p = os.path.join(out_dir, dir)
+            if os.path.isdir(p):
+                if 'result.log' in os.listdir(p):
+                    rp = os.path.join(p, 'result.log')
+                    logger.info(f'Find: {rp}')
+                    with open(rp, 'r') as f:
+                        final_f.write(f.read())
+                        final_f.write('\n')
+    logger.info(f'Final result saved at ' + os.path.join(out_dir, 'final_result.log'))