+++ datasets/derived_ards_dataset.py
@@ -37,11 +37,7 @@ def load_sepsis_patients(csv_path:str) -> tuple:
     return set(sepsis_dict.keys()), sepsis_dict
 
 class MIMICIV_ARDS_Dataset(MIMICIV_Core):
-    __name = 'mimic-iv-ards'
-
-    @classmethod
-    def name(cls):
-        return cls.__name
+    _name = 'mimic-iv-ards'
     
     def __init__(self):
         super().__init__(self.name())
@@ -88,7 +84,7 @@ class MIMICIV_ARDS_Dataset(MIMICIV_Core):
         if source == 'admission':
             admittime = ymdhms_converter(row.admittime)
             dischtime = ymdhms_converter(row.dischtime)
-            if dischtime < admittime:
+            if dischtime <= admittime:
                 return
             adm = Admission(
                 unique_id=int(row.hadm_id*1e8),
@@ -333,88 +329,3 @@ class MIMICIV_ARDS_Dataset(MIMICIV_Core):
             'dynamic_keys': dynamic_keys
         }
     
-    def make_report(self, version_name, params:dict):
-        # switch version
-        self.load_version(version_name)
-        self.mode('all')
-        out_path = os.path.join(self._paths['out_dir'], f'dataset_report_{version_name}.txt')
-        dist_dir = os.path.join(self._paths['out_dir'], 'report_dist')
-        dir_names = ['points', 'duration', 'frequency', 'value', 'from_sepsis', 'static_value']
-        tools.reinit_dir(dist_dir, build=True)
-        for name in dir_names:
-            os.makedirs(os.path.join(dist_dir, name))
-        logger.info('MIMIC-IV-ARDS: generating dataset report')
-        write_lines = []
-        if params['basic']:
-            # basic statistics
-            write_lines.append('='*10 + 'basic' + '='*10)
-            write_lines.append(f'Version: {version_name}')
-            write_lines.append(f'Static keys: {len(self.static_keys)}')
-            write_lines.append(f'Dynamic keys: {len(self.dynamic_keys)}')
-            write_lines.append(f'Subjects:{len(self)}')
-            write_lines.append(f'Static feature: {[self.fea_label(id) for id in self.static_keys]}')
-            write_lines.append(f'Dynamic feature: {[self.fea_label(id) for id in self.dynamic_keys]}')
-        if params['dynamic_dist']:
-            # dynamic feature explore
-            for id in tqdm(self.dynamic_keys, 'plot dynamic dist'):
-                fea_name = self.fea_label(id)
-                save_name = tools.remove_slash(str(fea_name))
-                write_lines.append('='*10 + f'{fea_name}({id})' + '='*10)
-                arr_points = []
-                arr_duration = []
-                arr_frequency = []
-                arr_avg_value = []
-                arr_from_sepsis_time = []
-                for s in self._subjects.values():
-                    for adm in s.admissions:
-                        if id in adm.keys():
-                            dur = adm[id][-1,1] - adm[id][0,1]
-                            sepsis_time = s.latest_static('sepsis_time', adm[id][0, 1])
-                            t_sep = adm[id][0, 1] + adm.admittime - sepsis_time
-                            arr_from_sepsis_time.append(t_sep)
-                            arr_points.append(adm[id].shape[0])
-                            arr_duration.append(dur)
-                            if dur > 1e-3: # TODO 只有一个点无法计算
-                                arr_frequency.append(arr_points[-1] / arr_duration[-1])
-                            else:
-                                arr_frequency.append(0)
-                            arr_avg_value.append(adm[id][:,0].mean())
-                arr_points, arr_duration, arr_frequency, arr_avg_value = \
-                    np.asarray(arr_points), np.asarray(arr_duration), np.asarray(arr_frequency), np.asarray(arr_avg_value)
-                arr_from_sepsis_time = np.asarray(arr_from_sepsis_time)
-                if np.size(arr_points) != 0:
-                    write_lines.append(f'average points per admission: {arr_points.mean():.3f}')
-                if np.size(arr_duration) != 0:
-                    write_lines.append(f'average duration(hour) per admission: {arr_duration.mean():.3f}')
-                if np.size(arr_frequency) != 0:
-                    write_lines.append(f'average frequency(point/hour) per admission: {arr_frequency.mean():.3f}')
-                if np.size(arr_avg_value) != 0:
-                    write_lines.append(f'average avg value per admission: {arr_avg_value.mean():.3f}')
-                # plot distribution
-                titles = ['points', 'duration', 'frequency', 'value', 'from_sepsis']
-                arrs = [arr_points, arr_duration, arr_frequency, arr_avg_value, arr_from_sepsis_time]
-                for title, arr in zip(titles, arrs):
-                    if np.size(arr) != 0:
-                        tools.plot_single_dist(
-                            data=arr, data_name=f'{title}: {fea_name}', 
-                            save_path=os.path.join(dist_dir, title, save_name + '.png'), discrete=False, adapt=True, bins=50)
-        if params['static_dist']:
-            # static feature explore
-            for id in tqdm(self.static_keys, 'generate static feature report'):
-                fea_name = self.fea_label(id)
-                save_name = tools.remove_slash(str(fea_name))
-                write_lines.append('='*10 + f'{fea_name}({id})' + '='*10)
-                idx = self.idx_dict[str(id)]
-                static_data = self.data[:, idx, 0]
-                write_lines.append(f'mean: {static_data.mean():.3f}')
-                write_lines.append(f'std: {static_data.std():.3f}')
-                write_lines.append(f'max: {np.max(static_data):.3f}')
-                write_lines.append(f'min: {np.min(static_data):.3f}')
-                tools.plot_single_dist(
-                    data=static_data, data_name=f'{fea_name}', 
-                    save_path=os.path.join(dist_dir, 'static_value', save_name + '.png'), discrete=False, adapt=True, bins=50)
-        # write report
-        with open(out_path, 'w', encoding='utf-8') as fp:
-            for line in write_lines:
-                fp.write(line + '\n')
-        logger.info(f'Report generated at {out_path}')