+++ models/__init__.py
@@ -1,1 +1,1 @@
-from .utils import *
+from ..tools.data_generator.data_generator import *